[indexes] move prover and verifier indexes in their own modules
