+++ kimchi/src/lib.rs
@@ -5,10 +5,11 @@ pub mod alphas;
 pub mod bench;
 pub mod circuits;
 pub mod error;
-pub mod index;
 pub mod plonk_sponge;
 pub mod prover;
+pub mod prover_index;
 pub mod verifier;
+pub mod verifier_index;
 
 #[cfg(test)]
 mod tests;
