+++ kimchi/src/tests/chacha.rs
@@ -4,8 +4,8 @@ use crate::{
         polynomials::chacha,
         wires::{Wire, COLUMNS},
     },
-    index::testing::new_index_for_test,
     prover::ProverProof,
+    prover_index::testing::new_index_for_test,
     verifier::batch_verify,
 };
 use array_init::array_init;
