+++ kimchi/src/tests/ec.rs
@@ -3,8 +3,8 @@ use crate::{
         gate::{CircuitGate, GateType},
         wires::*,
     },
-    index::testing::new_index_for_test,
     prover::ProverProof,
+    prover_index::testing::new_index_for_test,
     verifier::batch_verify,
 };
 use ark_ec::{AffineCurve, ProjectiveCurve};
